Refactor profile page JavaScript for modularity and better event handling
