+++ js/profile/redirectToHomePage.js
@@ -1,6 +1,6 @@
-const backend_url = 'http://localhost:3001'
-document.addEventListener('DOMContentLoaded', () => {
+import { isAuthenticated } from "./utils/auth.js";
 
+document.addEventListener('DOMContentLoaded', () => {
     // Get links to the specific sections of the homepage
     const links = document.querySelectorAll('.section')
     links.forEach(link => {
@@ -17,18 +17,4 @@ document.addEventListener('DOMContentLoaded', () => {
     });
 })
 
-// Function to check if the user is authenticated
-function isAuthenticated() {
-    return fetch(`${backend_url}/auth/status`, {
-        method: 'GET',
-        credentials: 'include'
-    })
-        .then(response => response.json())
-        .then(data => data.isAuthenticated)
-        .catch(error => {
-            console.error('Error:', error);
-            return false;
-        });
-}
 
-export { isAuthenticated }
